requirements.txt
flask
flask-sqlalchemy
flask-smorest
marshmallow
python-dotenv
SQLite